requirements.txt
Flask==3.0.2
Werkzeug==3.0.1

tensorflow==2.20.0

Pillow
gunicorn==21.2.0
python-dotenv